requirements.txt
humanize==3.4.1
python-telegram-bot==13.4.1
requests==2.25.1
youtube-dl==2021.4.17
